Move model from sec to finra
