+++ openbb_platform/providers/finra/openbb_finra/__init__.py
@@ -1,4 +1,5 @@
 """FINRA provider module."""
+from openbb_finra.models.equity_short_interest import FinraShortInterestFetcher
 from openbb_finra.models.otc_aggregate import FinraOTCAggregateFetcher
 from openbb_provider.abstract.provider import Provider
 
@@ -7,5 +8,8 @@ finra_provider = Provider(
     website="https://finra.org",
     description="Financial Industry Regulatory Authority.",
     required_credentials=None,
-    fetcher_dict={"OTCAggregate": FinraOTCAggregateFetcher},
+    fetcher_dict={
+        "OTCAggregate": FinraOTCAggregateFetcher,
+        "EquityShortInterest": FinraShortInterestFetcher,
+    },
 )
