+++ openbb_platform/providers/sec/openbb_sec/__init__.py
@@ -2,7 +2,6 @@
 from openbb_provider.abstract.provider import Provider
 from openbb_sec.models.cik_map import SecCikMapFetcher
 from openbb_sec.models.company_filings import SecCompanyFilingsFetcher
-from openbb_sec.models.equity_short_interest import SecShortInterestFetcher
 from openbb_sec.models.institutions_search import SecInstitutionsSearchFetcher
 from openbb_sec.models.rss_litigation import SecRssLitigationFetcher
 from openbb_sec.models.schema_files import SecSchemaFilesFetcher
@@ -27,6 +26,5 @@ sec_provider = Provider(
         "StockSearch": SecStockSearchFetcher,
         "SymbolMap": SecSymbolMapFetcher,
         "RssLitigation": SecRssLitigationFetcher,
-        "EquityShortInterest": SecShortInterestFetcher,
     },
 )
